+++ src/main/java/com/kakaotech/team14backend/point/appliation/PointService.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.outer.point.service;
+package com.kakaotech.team14backend.point.appliation;
 
 import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import com.kakaotech.team14backend.inner.member.model.Member;
